+++ src/internal/internal.c
@@ -56,6 +56,7 @@ void rlang_init_internal(r_obj* ns) {
   rlang_init_cnd(ns);
   rlang_init_cnd_handlers(ns);
   rlang_init_dots(ns);
+  rlang_init_env();
   rlang_init_expr_interp();
   rlang_init_eval_tidy();
   rlang_init_fn();
